Properly exit Bedrock layer tasks
