+++ test-level/db/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000709
+MANIFEST-000719
